backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI / Aiogram"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "alembic>=1.16.1",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "fastapi-cache2>=0.2.2",
    "loguru>=0.7.3",
    "msgpack>=1.1.0",
    "orjson>=3.10.18",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "redis[hiredis]>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
]
